requirements.txt
flask~=3.1.0
sqlalchemy~=2.0.38
#python-telegram-bot~=21.0.1
#requests
qrcode~=8.0
werkzeug~=3.1.3
pillow==11.1.0

Flask-Migrate~=3.1.0
itsdangerous~=2.2.0
pyperclip~=1.9.0
alembic~=1.7.7

PyJWT~=2.10.1
Flask-JWT-Extended~=4.7.1
python-telegram-bot~=21.0.1
opencv-python~=4.11.0.86
numpy~=2.2.3
cloudinary~=1.42.2
requests~=2.32.3
marshmallow~=3.26.1
Flask-Limiter~=3.12